Remove unused files
The `doc`, `locale` and `test` directories only contain placeholder
files, this makes it harder for a new maintainer to understand the
structure of the project so I simply removed them. Removed the
`package.json.backup` as well, it's a git repository, no need to keep
`.backup` files.

Conflicts:
	test/test-main.js

Keeping the test files for the main module