updating the urlbutton and the match function
Now fully supporting the tristate match of invalid urls, valid urls
without ratings, and valid urls with ratings.
Our url button responds correctly to each state as shown in the tests.
Created a test for the match function to ensure our tristate is
returned correctly.
Updated the notifications and panel to the new match function.